Seperate graphql test to schema and mutation for Project + AF
