+++ .gitignore
@@ -34,6 +34,7 @@ generated/
 *.pyc
 LOCAL_DATA/
 .idea/
+media-temp/
 
 # Generated by docker on up (only for mypy)
 .mypy-dep/
